create delay.rs
